Fix building of OSNMA tests
